passai-lite/requirements.txt
PyQt6==6.6.1
PyQt6-Qt6==6.6.1
PyQt6-sip==13.6.0
cryptography==41.0.7
argon2-cffi==23.1.0
Pillow==10.1.0
requests==2.31.0
